split geoaccessor to script
